Move component and redux files out of pages
