+++ frontend/pages/_app.tsx
@@ -1,11 +1,11 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
-import MyHead from './components/Layout/MyHead'
-import Header from './components/Layout/Header'
-import Footer from './components/Layout/Footer'
-import Body from './components/Layout/Body'
+import MyHead from '../components/Layout/MyHead'
+import Header from '../components/Layout/Header'
+import Footer from '../components/Layout/Footer'
+import Body from '../components/Layout/Body'
 import { Provider } from 'react-redux'
-import { store } from './redux/store/store'
+import { store } from '../redux/store/store'
 import { ToastContainer } from 'react-toastify'
 import "react-toastify/dist/ReactToastify.css";
