+++ frontend/pages/index.tsx
@@ -1,8 +1,8 @@
-import Form from './components/Form/Form'
+import Form from '../components/Form/Form'
 import { useEffect } from 'react';
-import { useAppSelector, useAppDispatch } from './redux/hooks/hooks';
+import { useAppSelector, useAppDispatch } from '../redux/hooks/hooks';
 import { useRouter } from 'next/router';
-import { checkUser } from './redux/slice/authSlice';
+import { checkUser } from '../redux/slice/authSlice';
 import { toast } from 'react-toastify';
 export default function Home() {
